--- bin/tests/node/AsynchronousWorkflow.js
@@ -1,8 +1,0 @@
-require("../../../init");
-require("../../../client");
-
-const AsynchronousWorkflow = require(paths.testsNode + "AsynchronousWorkflow");
-
-new AsynchronousWorkflow().dispatch().catch(err => {
-  console.error(err);
-});
