--- bin/tests/node/ErrorWorkflow.js
@@ -1,8 +1,0 @@
-require("../../../init");
-require("../../../client");
-
-const ErrorWorkflow = require(paths.testsNode + "ErrorWorkflow");
-
-new ErrorWorkflow().dispatch().catch(err => {
-  console.error(err);
-});
