--- bin/tests/node/SequentialWorkflow.js
@@ -1,8 +1,0 @@
-require("../../../init");
-require("../../../client");
-
-const SequentialWorkflow = require(paths.testsNode + "SequentialWorkflow");
-
-new SequentialWorkflow().dispatch().catch(err => {
-  console.error(err);
-});
