--- bin/tests/node/WaitWorkflow.js
@@ -1,8 +1,0 @@
-require("../../../init");
-require("../../../client");
-
-const WaitWorkflow = require(paths.testsNode + "WaitWorkflow");
-
-new WaitWorkflow().dispatch().catch(err => {
-  console.error(err);
-});
