+++ boot.js
@@ -7,3 +7,10 @@ require("./tests/node/ErrorWorkflow");
 require("./tests/node/EventWorkflow");
 require("./tests/node/WaitEventWorkflow");
 require("./tests/node/WaitWorkflow");
+
+require("./tests/SequentialWorkflow");
+require("./tests/AsynchronousWorkflow");
+require("./tests/ErrorWorkflow");
+require("./tests/EventWorkflow");
+require("./tests/WaitEventWorkflow");
+require("./tests/WaitWorkflow");
