+++ output/expected/node/AsynchronousWorkflow
@@ -1,6 +1,6 @@
-Task B starts
 Task A starts
 Task C starts
+Task B starts
 Task A ends
 Task B ends
 Task C ends
