+++ output/expected/node/ErrorWorkflow
@@ -1,3 +1,3 @@
-Task E starts
 Task A starts
+Task E starts
 Task A ends
