--- undefinedSequentialWorkflow
@@ -1,6 +1,0 @@
-Task A starts
-Task A ends
-Task C starts
-Task C ends
-Task D starts
-Task D ends
