--- undefinedWaitEventWorkflow
@@ -1,2 +1,0 @@
-Task A starts
-Task A ends
