--- undefinedWaitWorkflow
@@ -1,4 +1,0 @@
-Task A starts
-Task A ends
-Task B starts
-Task B ends
